Move the Elixir version requirement check to a better place
