+++ lib/mix/lib/mix/tasks/loadpaths.ex
@@ -21,6 +21,18 @@ defmodule Mix.Tasks.Loadpaths do
       Mix.Task.run "deps.loadpaths", args
     end
 
+    unless opts[:no_elixir_version_check] do
+      config = Mix.project
+
+      if requirement = config[:elixir] do
+        unless Mix.Version.match?(System.version, requirement) do
+          raise Mix.SystemVersionError, target: config[:app] || Mix.Project.get,
+                                        expected: requirement,
+                                        actual: System.version
+        end
+      end
+    end
+
     Enum.each Mix.Project.load_paths, Code.prepend_path(&1)
   end
 end
