+++ lib/mix/test/mix/tasks/app.start_test.exs
@@ -9,6 +9,12 @@ defmodule Mix.Tasks.App.StartTest do
     end
   end
 
+  defmodule WrongElixirProject do
+    def project do
+      [ app: :error, version: "0.1.0", elixir: "~> 0.8.1" ]
+    end
+  end
+
   test "compile and starts the project" do
     Mix.Project.push CustomApp
 
@@ -24,6 +30,32 @@ defmodule Mix.Tasks.App.StartTest do
 
       Mix.Tasks.App.Start.run []
       assert List.keyfind(:application.loaded_applications, :app_start_sample, 0)
+    end
+  after
+    purge [A, B, C]
+    Mix.Project.pop
+  end
+
+  test "validates the Elixir version requirement" do
+    Mix.Project.push WrongElixirProject
+
+    in_fixture "no_mixfile", fn ->
+      error = assert_raise Mix.SystemVersionError, fn ->
+        Mix.Tasks.App.Start.run ["--no-start"]
+      end
+
+      assert error.message =~ %r/ to run :error /
+    end
+  after
+    purge [A, B, C]
+    Mix.Project.pop
+  end
+
+  test "does not validate the Elixir version requirement when disabled" do
+    Mix.Project.push WrongElixirProject
+
+    in_fixture "no_mixfile", fn ->
+      Mix.Tasks.App.Start.run ["--no-start", "--no-elixir-version-check"]
     end
   after
     purge [A, B, C]
